Fan improvements (not worth the effort at all)
but made code newer and fixed some bugs that didnt really make bad things.

+++ Repetier/Printer.h
@@ -1218,8 +1218,8 @@ public:
     static INLINE int getFanSpeed(bool percent = false)
     {
         if(!percent) return (int)pwm_pos[NUM_EXTRUDER+2]; //int
-        if(!pwm_pos[NUM_EXTRUDER+2]) return (int)0; //%
-        if(pwm_pos[NUM_EXTRUDER+2] <= 3) return (int)1; //%
+        if(!pwm_pos[NUM_EXTRUDER+2]) return 0; //%
+        if(pwm_pos[NUM_EXTRUDER+2] <= 3) return 1; //%
         return (int)(pwm_pos[NUM_EXTRUDER+2]*100/255); //%
     } // getFanSpeed
